tests

+++ meldingen/asset.py
@@ -1,4 +1,3 @@
-from meldingen.models import Melding
 from meldingen.repositories import MeldingRepository
 
 
@@ -8,5 +7,10 @@ class AssetPurger:
     def __init__(self, repository: MeldingRepository):
         self._repository = repository
 
-    async def __call__(self, melding: Melding) -> None:
+    async def __call__(self, melding_id: int) -> None:
+        melding = await self._repository.retrieve(melding_id)
+
+        if melding is None:
+            return
+
         await self._repository.delete_assets_from_melding(melding)
